Perf: Implement init helper for both Dashboard and Popup
